# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Azubi Commerce API",
        "description": "Azubi Commerce API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "local server"
        },
        {
            "url": "https://example.com",
            "description": "production server"
        }
    ],
    "paths": {
        "/create": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Save user information and generate token",
                "description": "Saves user to the database",
                "operationId": "02092ff663977235ddd651857ce2c8ec",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "User Name"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "email",
                            "format": "string",
                            "example": "johnson@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of User",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "dfs@.drL;s"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created Successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "xxxxxxxxxxxxxxxxxxxxxx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": "name field cannot be empty"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Log user in",
                "description": "Saves user to the database",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "email",
                            "format": "string",
                            "example": "johnson@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of User",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "dfs@.drL;s"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User logged in Successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "xxxxxxxxxxxxxxxxxxxxxx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": "email field cannot be empty"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "brearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "header"
            }
        }
    },
...
}